signVerify/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
keras==2.7.0
matplotlib==3.4.2
Pillow==8.4.0
scikit-image==0.18.3
scikit-learn
tensorflow==2.7.0
gunicorn==20.0.4
